requirements.txt
Pillow==2.6.1
vdf==1.5
vpk==0.7
